+++ tests/common/snappi/traffic_generation.py
