+++ tests/snappi/files/helper.py
