+++ tests/snappi/variables.py
